+++ src/Lm/OnnxRecurrentLanguageModel.cc
@@ -1,6 +1,9 @@
 #include "OnnxRecurrentLanguageModel.hh"
 
+#include "AbstractStateManager.hh"
 #include "OnnxLstmStateManager.hh"
+#include "OnnxTransformerStateManager.hh"
+
 #include "OnnxNceSoftmaxAdapter.hh"
 #include "OnnxPassthroughSoftmaxAdapter.hh"
 
@@ -8,9 +11,30 @@ namespace {
 
 std::vector<Onnx::IOSpecification> getIOSpec(int64_t num_classes) {
     return std::vector<Onnx::IOSpecification>({
-            Onnx::IOSpecification{"word", Onnx::IODirection::INPUT, false, {Onnx::ValueType::TENSOR}, {Onnx::ValueDataType::INT32}, {{-1, -1}}},
-            Onnx::IOSpecification{"word-length", Onnx::IODirection::INPUT, false, {Onnx::ValueType::TENSOR}, {Onnx::ValueDataType::INT32}, {{-1}}},
-            Onnx::IOSpecification{"nn-output", Onnx::IODirection::OUTPUT, false, {Onnx::ValueType::TENSOR}, {Onnx::ValueDataType::FLOAT}, {{-1, -1, num_classes}}},
+            Onnx::IOSpecification{"word",
+                                  Onnx::IODirection::INPUT,
+                                  false,
+                                  {Onnx::ValueType::TENSOR},
+                                  {Onnx::ValueDataType::INT32},
+                                  {{-1, -1}}},
+            Onnx::IOSpecification{"word-length",
+                                  Onnx::IODirection::INPUT,
+                                  false,
+                                  {Onnx::ValueType::TENSOR},
+                                  {Onnx::ValueDataType::INT32},
+                                  {{-1}}},
+            Onnx::IOSpecification{"prefix-length",
+                                  Onnx::IODirection::INPUT,
+                                  true,
+                                  {Onnx::ValueType::TENSOR},
+                                  {Onnx::ValueDataType::INT32},
+                                  {{-1}}},
+            Onnx::IOSpecification{"nn-output",
+                                  Onnx::IODirection::OUTPUT,
+                                  false,
+                                  {Onnx::ValueType::TENSOR},
+                                  {Onnx::ValueDataType::FLOAT},
+                                  {{-1, -1, num_classes}}},
     });
 }
 
@@ -23,27 +47,37 @@ enum OnnxStateManagerType {
     OnnxTransformerStateManagerType,
     OnnxTransformerStateManager16BitType,
     OnnxTransformerStateManager8BitType,
-    OnnxTransformerStateManagerWithCommonPrefixType,
-    OnnxTransformerStateManagerWithCommonPrefix16BitType,
-    OnnxTransformerStateManagerWithCommonPrefix8BitType,
 };
 
 const Core::Choice stateManagerTypeChoice(
         "lstm", OnnxLstmStateManagerType,
+        "transformer", OnnxTransformerStateManagerType,
+        "transformer-16bit", OnnxTransformerStateManager16BitType,
+        "transformer-8bit", OnnxTransformerStateManager8BitType,
         Core::Choice::endMark());
 
-const Core::ParameterChoice stateManagerTypeParam(
-        "type", &stateManagerTypeChoice,
-        "type of the state manager",
-        OnnxLstmStateManagerType);
+const Core::ParameterChoice stateManagerTypeParam("type",
+                                                  &stateManagerTypeChoice,
+                                                  "type of the state manager",
+                                                  OnnxLstmStateManagerType);
 
 std::unique_ptr<OnnxStateManager> createOnnxStateManager(Core::Configuration const& config) {
-    OnnxStateManager* res = nullptr;
     switch (stateManagerTypeParam(config)) {
-        case OnnxLstmStateManagerType: res = new Lm::OnnxLstmStateManager(config); break;
-        default: defect();
+        case OnnxLstmStateManagerType:
+            return std::unique_ptr<OnnxStateManager>(
+                    new Lm::OnnxLstmStateManager(config));
+        case OnnxTransformerStateManagerType:
+            return std::unique_ptr<OnnxStateManager>(
+                    new Lm::OnnxTransformerStateManager<float>(config));
+        case OnnxTransformerStateManager16BitType:
+            return std::unique_ptr<OnnxStateManager>(
+                    new Lm::OnnxTransformerStateManager<int16_t>(config));
+        case OnnxTransformerStateManager8BitType:
+            return std::unique_ptr<OnnxStateManager>(
+                    new Lm::OnnxTransformerStateManager<int8_t>(config));
+        default:
+            defect();
     }
-    return std::unique_ptr<OnnxStateManager>(res);
 }
 
 enum OnnxSoftmaxAdapterType {
@@ -51,21 +85,24 @@ enum OnnxSoftmaxAdapterType {
     OnnxNceSoftmaxAdapterType,
 };
 
-const Core::Choice softmaxAdapterTypeChoice(
-        "passthrough", OnnxPassthroughSoftmaxAdapterType,
-        "nce", OnnxNceSoftmaxAdapterType,
-        Core::Choice::endMark());
+const Core::Choice softmaxAdapterTypeChoice("passthrough",
+                                            OnnxPassthroughSoftmaxAdapterType,
+                                            "nce", OnnxNceSoftmaxAdapterType,
+                                            Core::Choice::endMark());
 
-const Core::ParameterChoice softmaxAdapterTypeParam(
-        "type", &softmaxAdapterTypeChoice,
-        "type of the softmax adapter",
-        OnnxPassthroughSoftmaxAdapterType);
+const Core::ParameterChoice softmaxAdapterTypeParam("type",
+                                                    &softmaxAdapterTypeChoice,
+                                                    "type of the softmax adapter",
+                                                    OnnxPassthroughSoftmaxAdapterType);
 
 std::unique_ptr<OnnxSoftmaxAdapter> createOnnxSoftmaxAdapter(Core::Configuration const& config) {
     switch (softmaxAdapterTypeParam(config)) {
-        case OnnxPassthroughSoftmaxAdapterType: return std::unique_ptr<OnnxSoftmaxAdapter>(new Lm::OnnxPassthroughSoftmaxAdapter(config));
-        case OnnxNceSoftmaxAdapterType: return std::unique_ptr<OnnxSoftmaxAdapter>(new Lm::OnnxNceSoftmaxAdapter(config));
-        default: defect();
+        case OnnxPassthroughSoftmaxAdapterType:
+            return std::unique_ptr<OnnxSoftmaxAdapter>(new Lm::OnnxPassthroughSoftmaxAdapter(config));
+        case OnnxNceSoftmaxAdapterType:
+            return std::unique_ptr<OnnxSoftmaxAdapter>(new Lm::OnnxNceSoftmaxAdapter(config));
+        default:
+            defect();
     }
 }
 
@@ -85,24 +122,35 @@ OnnxRecurrentLanguageModel::OnnxRecurrentLanguageModel(Core::Configuration const
     validator_.validate(ioSpec_, mapping_, session_);
 }
 
-void OnnxRecurrentLanguageModel::setState(std::vector<std::pair<std::string, Onnx::Value>> const& inputs, std::vector<std::string> const& targets) const {
-}
+void OnnxRecurrentLanguageModel::setState(std::vector<std::pair<std::string, Onnx::Value>> const& inputs,
+                                          std::vector<std::string> const&                         targets) const {}
 
-void OnnxRecurrentLanguageModel::extendInputs(std::vector<std::pair<std::string, Onnx::Value>>& inputs, Math::FastMatrix<s32> const& words, Math::FastVector<s32> const& word_lengths, std::vector<s32> const& state_lengths) const {
+void OnnxRecurrentLanguageModel::extendInputs(std::vector<std::pair<std::string, Onnx::Value>>& inputs,
+                                              Math::FastMatrix<s32> const&                      words,
+                                              Math::FastVector<s32> const&                      word_lengths,
+                                              std::vector<s32> const&                           state_lengths) const {
     inputs.emplace_back(mapping_.getOnnxName("word"), Onnx::Value::create(words));
     inputs.emplace_back(mapping_.getOnnxName("word-length"), Onnx::Value::create(word_lengths));
+    auto prefix_length_name = mapping_.getOnnxName("prefix-length");
+    if (prefix_length_name != "") {
+        inputs.emplace_back(prefix_length_name, Onnx::Value::create(state_lengths));
+    }
 }
 
 void OnnxRecurrentLanguageModel::extendTargets(std::vector<std::string>& targets) const {
     targets.emplace(targets.begin(), mapping_.getOnnxName("nn-output"));
 }
 
-void OnnxRecurrentLanguageModel::getOutputs(std::vector<std::pair<std::string, Onnx::Value>>& inputs, std::vector<Onnx::Value>& outputs, std::vector<std::string> const& targets) const {
+void OnnxRecurrentLanguageModel::getOutputs(std::vector<std::pair<std::string, Onnx::Value>>& inputs,
+                                            std::vector<Onnx::Value>&                         outputs,
+                                            std::vector<std::string> const&                   targets) const {
     session_.run(std::move(inputs), targets, outputs);
 }
 
 std::vector<Onnx::Value> OnnxRecurrentLanguageModel::fetchStates(std::vector<Onnx::Value>& outputs) const {
-    std::vector<Onnx::Value> state_vars(std::make_move_iterator(outputs.begin() + 1), std::make_move_iterator(outputs.end()));
+    std::vector<Onnx::Value> state_vars(
+            std::make_move_iterator(outputs.begin() + 1),
+            std::make_move_iterator(outputs.end()));
     return state_vars;
 }
